+++ src/Author/MyWorks/ContentViews/ColumnDisplay.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
-import EditingButtons from "./EditingButtons";
+import EditingButtons from "../EditingButtons";
 
 const useStyles = makeStyles(theme => ({
   grid: {
